spike(postgres): catálogo, fechas y MIN/MAX — de 69% a 74% de la suite
Los 15 sitios de PRAGMA table_info y sqlite_master se resuelven en UN lugar
(el shim) y no en 15: se revisó uno por uno y ninguno de los 9 de main.py
necesita más que el NOMBRE de la columna. Verificado contra el esquema real:
58 tablas, 0 diferencias de columnas ni de orden.

Con caché. La primera versión andaba pero BAJÓ el resultado (1960→1734): en
SQLite preguntar las columnas es gratis, en Postgres es una consulta, y
_table_cols() se llama ~20 veces por request. Era un problema de producción
disfrazado de problema de tests: contra Supabase serían ~20 viajes por pantalla.
El caché se invalida con cualquier DDL — sin eso, init_db() creería que una
columna ya está y se la saltearía.

Fechas con modificador: la categoría que más callada falla. `datetime('now')`
ya se traducía; `datetime('now','-7 days')` no, y como varios de esos sitios
están dentro de un try/except, el error se tragaba y la función devolvía un
default. El usuario veía un número equivocado, no un error. Eran 15 sitios, no
los 12 contados a mano: faltaban date(MIN(x),'+7 days') —el paréntesis cortaba
el match— y dos date(columna) que Postgres ACEPTA pero devolviendo un objeto
fecha en vez del texto 'YYYY-MM-DD'. Probado corriendo la misma expresión en
las dos bases: 15 de 15 idénticos, con bisiesto y fines de mes.

MIN(a,b)/MAX(a,b) → LEAST/GREATEST. En Postgres son agregados de un argumento:
no dan otro resultado, no existen. Se cuentan paréntesis y no se usa un regex,
porque confundir MIN(COALESCE(a,0)) con dos argumentos cambiaría una consulta
en silencio.

Fechas mal formadas: SQLite daba NULL y la fila se salteaba sola; Postgres
corta con error. Se filtran ANTES del CAST con un LIKE que se comporta igual en
los dos motores. Uno de los 3 sitios no tenía ni el filtro de nulos y agrupaba
por año/mes: una sola fecha rota tiraba abajo el agregado completo.

FCI: price REAL → DOUBLE PRECISION. Esa tabla se crea fuera del esquema, así
que no pasa por el traductor de tipos; en Postgres REAL son 4 bytes y redondeaba
los precios. En SQLite es idéntico.

Medido: 1960→2127 tests (69%→74%). Errores de fecha 215→0. Las AssertionError
bajaron de 42 a 27 SIN tocar un test: eran cascada, no diferencias semánticas.
Regresión en SQLite: las mismas 46 fallas preexistentes, ni una más.

Los 556 timeouts que quedan NO son de la app: los 58 archivos de test comparten
UNA base y se aíslan con DROP SCHEMA, que se cuelga si un test anterior dejó una
conexión abierta. Verificado con un experimento (se colgó 8s y murió).

+++ backend/pricing/fci.py
@@ -187,7 +187,13 @@ def ensure_tables(conn):
         );
         CREATE TABLE IF NOT EXISTS fci_prices (
             symbol     TEXT PRIMARY KEY,
-            price      REAL,
+            -- DOUBLE PRECISION y no REAL: estas dos tablas se crean acá y no en
+            -- el esquema, así que no pasan por el traductor que convierte los
+            -- tipos. En SQLite `REAL` son 8 bytes, pero en Postgres son 4 —unos
+            -- 7 dígitos significativos— y esto es una columna de PRECIOS: un vcp
+            -- como 1234567.89 se guardaría redondeado. `DOUBLE PRECISION` vale
+            -- en los dos motores y en SQLite es exactamente lo mismo que REAL.
+            price      DOUBLE PRECISION,
             moneda     TEXT,
             as_of_date TEXT,
             fetched_at TEXT
